Upgrade to Lucene 4.10.4 keeping the LUCENE_43 index format.

+++ src/main/java/io/jitter/core/selection/ShardStats.java
@@ -13,7 +13,7 @@ public class ShardStats {
         maxSize = Collections.max(sizes.values());
         totalDocs = 0;
         for (Integer sz : sizes.values()) {
-            totalDocs += sz.intValue();
+            totalDocs += sz;
         }
     }
